🚚 Move projector and notification infra in the right place
