+++ packages/applications/notifications/src/infrastructure/sendEmail.ts
