+++ packages/applications/notifications/src/infrastructure/sendEmailMode.ts
