+++ packages/applications/notifications/src/lauréat/abandon.notification.ts
@@ -1,5 +1,5 @@
 import { Message, MessageHandler, mediator } from 'mediateur';
-import { sendEmail } from '@potentiel-libraries/email-sender';
+import { sendEmail } from '../infrastructure/sendEmail';
 import { IdentifiantProjet } from '@potentiel-domain/common';
 import { Option } from '@potentiel-libraries/monads';
 import { Event } from '@potentiel-infrastructure/pg-event-sourcing';
