+++ packages/applications/projectors/src/infrastructure/createProjection.ts
