+++ packages/applications/projectors/src/infrastructure/removeProjection.ts
