+++ packages/applications/projectors/src/infrastructure/removeProjectionByCategory.ts
