+++ packages/applications/projectors/src/infrastructure/updateProjection.ts
