--- packages/libraries/email-sender/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@potentiel-libraries/email-sender",
-  "version": "0.0.0",
-  "description": "Email sender package to send emails",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "private": true,
-  "scripts": {
-    "build": "tsc"
-  },
-  "devDependencies": {
-    "@potentiel-config/tsconfig": "*",
-    "@types/node-mailjet": "^3.3.9"
-  },
-  "dependencies": {
-    "@potentiel-libraries/monitoring": "*",
-    "node-mailjet": "^6.0.4"
-  }
-}
