--- packages/libraries/email-sender/src/index.ts
@@ -1,1 +1,0 @@
-export { sendEmail } from './sendEmail';
