--- packages/libraries/email-sender/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "@potentiel-config/tsconfig/tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "./src"
-  }
-}
